Revert "Deployment script"
