--- frontend/jsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-      "module": "commonjs",
-      "target": "es2016",
-      "jsx": "preserve"
-  },
-  "exclude": [
-      "node_modules",
-      "**/node_modules/*"
-  ]
-}
