+++ frontend/src/App.css
@@ -236,14 +236,13 @@ nav {
   padding: 1.2rem 0.1rem;
   display: flex;
   flex-direction: row;
-  justify-content: space-between;
-  /* background: linear-gradient(135deg, #fc575e, #f7b42c); */
+  justify-content: flex-end;
+  background: linear-gradient(135deg, #fc575e, #f7b42c);
   position: sticky;
   overflow: hidden;
   top: 0;
   width: 100%;
   z-index: 100;
-  background-color: #fff;
 }
 
 .NavMenu {
@@ -255,7 +254,7 @@ nav {
 
 .nav-link {
   display: flex;
-  color: rgba(37, 42, 55, 255);
+  color: #fff;
   align-items: center;
   text-decoration: none;
   padding: 0 1.5rem;
@@ -272,126 +271,10 @@ nav {
   width: 6.5rem;
   height: 2rem;
   padding: 1.5rem;
-  border: 2px solid hsl(0, 0%, 85%);
-  background: rgba(37, 42, 55, 255);
-  color: #fff;
+  border: 2px solid rgba(233, 156, 14, 0.479);
+  background: rgba(192, 129, 11, 0.479);
   border-radius: 10px;
   box-sizing: border-box;
-}
-
-.background {
-  background: rgba(37, 42, 55, 255);
-}
-
-.brand-logo {
-  margin-left: 11%;
-  height: 4.5rem;
-  margin-top: -1rem;
-}
-
-.landing-page {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: 5rem 10rem 0 11rem;
-}
-
-.active-btn {
-  background-color: rgb(225, 82, 42);
-  padding: 1rem 1.4rem;
-  color: #fff;
-  cursor: pointer;
-  border-radius: 10px;
-  border: 1px solid rgb(225, 82, 42);
-  margin-top: 1.5rem;
-  margin-bottom: 1rem;
-  margin-right: 1.2rem;
-  font-size: 14px;
-  font-weight: 800;
-  text-align: center;
-  transition: all 0.5s;
-}
-
-.active-btn::after {
-  content: "\00bb";
-  margin-left: 0.4rem;
-}
-
-/* .landing-page-text {
-  width: 100%;
-} */
-
-.landing-page-text-main {
-  color: #fff;
-  font-size: 50px;
-  font-weight: 900;
-  margin-bottom: 2rem;
-  margin-top: 1rem;
-}
-
-.landing-page-text-sub {
-  color: hsl(0, 0%, 85%);
-  font-size: 18px;
-  margin-bottom: 2rem;
-}
-
-.landing-page-image {
-  width: 100%;
-}
-
-.landing-page-button {
-  padding: 1rem;
-  color: #fff;
-  cursor: pointer;
-  border-radius: 10px;
-  border: 1px solid #fff;
-  margin-top: 1.5rem;
-  margin-bottom: 1rem;
-  margin-right: 1.2rem;
-  font-size: 14px;
-  font-weight: 800;
-  /* line-height: 1.5em; */
-  background-color: transparent;
-  text-align: center;
-  transition: all 0.5s;
-}
-
-.landing-page-button:hover {
-  background-color: rgb(225, 82, 42);
-  border: 1px solid rgb(225, 82, 42);
-  cursor: pointer;
-  display: inline-block;
-  position: relative;
-  transition: 0.5s;
-}
-
-.landing-page-button span {
-  cursor: pointer;
-  display: inline-block;
-  position: relative;
-  transition: 0.5s;
-}
-
-.landing-page-button span::after {
-  content: "\00bb";
-  position: absolute;
-  opacity: 0;
-  top: 0;
-  right: -20px;
-  transition: 0.5s;
-}
-
-.landing-page-button:hover span {
-  padding-right: 0.9rem;
-}
-
-.landing-page-button:hover span::after {
-  opacity: 1;
-  right: 0;
-}
-
-.bg-image {
-  margin-left: 20%;
 }
 
 #file-upload {
@@ -956,6 +839,7 @@ ul.meal-tab-nav li.active {
   display: flex;
   justify-content: flex-start;
   align-items: flex-end;
+  
 }
 
 #missing {
