--- frontend/src/components/MenuComponents/user-menu.component.js
@@ -1,16 +1,0 @@
-import "../../App.css";
-import MealTabs from "../../pages/meal-tabs";
-import NavBar from "../admin-homepage.component";
-
-function UserMenu() {
-  return (
-    <>
-      <NavBar />
-      <div className="admin-menu">
-        <MealTabs />
-      </div>
-    </>
-  );
-}
-
-export default UserMenu;
